feat(array): Allow ArrayEach to return error that breaks loop
Updated the function signature for the `ArrayEach` callback to remove
the pointless `error` parameter (`ArrayEach` will never invoke callback
if error is not nil) and added a `error` return type to the callback
signature.

If the callback returns a error the loop is broken and the callback
error and current offset is return.